chore: cover self-referencing api with test
closes https://github.com/sikanhe/gqtx/issues/37
